Move ingest methods to top-level
